+++ app/project/bootstrappers/http/views/ViewBootstrapper.php
@@ -4,10 +4,9 @@
  */
 namespace Project\Bootstrappers\HTTP\Views;
 
-use Opulence\Files\FileSystem;
 use Opulence\Framework\Bootstrappers\HTTP\Views\ViewBootstrapper as BaseBootstrapper;
 use Opulence\IoC\IContainer;
-use Opulence\Views\Caching\Cache;
+use Opulence\Views\Caching\FileCache;
 use Opulence\Views\Caching\ICache;
 
 class ViewBootstrapper extends BaseBootstrapper
@@ -21,11 +20,9 @@ class ViewBootstrapper extends BaseBootstrapper
      */
     protected function getViewCache(IContainer $container)
     {
-        $fileSystem = $container->makeShared(FileSystem::class);
         $cacheConfig = require_once "{$this->paths["configs.http"]}/views.php";
 
-        return new Cache(
-            $fileSystem,
+        return new FileCache(
             $this->paths["views.compiled"],
             $cacheConfig["cache.lifetime"],
             $cacheConfig["gc.chance"],